Rerun branch_optimizations examples on Ryzen and document results
Adds branch_free_ryzen.md, lut_branch_ryzen.md, and loop_unswitch_ryzen.md
comparing measurements on AMD Ryzen AI 9 HX 370 against the existing Apple
M4 baselines, including a breakdown of how each example's speedup shifts
across -O0..-O3 (notably GCC's vectorizer bypassing -fno-if-conversion at
-O2+, and lut_branch's speedup inverting at -O3). Also applies the same
(long long) cast fix to the three .cpp files' bench() helper needed to
compile under GCC/libstdc++ on Linux, matching the fix already used in
strength_reduction.cpp.

+++ branch_optimizations/branch_free.cpp
@@ -103,7 +103,7 @@ long long bench(Func f) {
         auto t0 = high_resolution_clock::now();
         f();
         auto t1 = high_resolution_clock::now();
-        best = min(best, duration_cast<milliseconds>(t1 - t0).count());
+        best = min(best, (long long)duration_cast<milliseconds>(t1 - t0).count());
     }
     return best;
 }
